--- src/main/kotlin/happyNumber/HappyNumber.kt
@@ -1,16 +1,0 @@
-package happyNumber
-
-class HappyNumber {
-    fun isHappy(n: Int): Boolean {
-        var number = n
-        while (number != 1 && number != 4) {
-            var sum = 0
-            while (number > 0) {
-                sum += (number % 10) * (number % 10)
-                number /= 10
-            }
-            number = sum
-        }
-        return number == 1
-    }
-}
